requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.4

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0
